Whatsapp chat analyser for analsing chats within few minutes
In olden days’ there is no analysis for whatsapp chat. If someone wants to analyze there is no CSV file to analyze. WhatsApp Application provide export txt file which is in raw format. It is very complicated for analysis. So we have to forget that system and switch over to the WhatsApp Chat Analyzer.